initial work on 'hydra'; a new database
Hydra is the name chosen for a new compiler-internal database, to be
used as a replacement for the existing solution in 'src/db.rs'. The file
'src/db_new.rs' exists mostly as a testing place to ensure that 'hydra'
does actually work as intended.

There's some high-level restructuring here in order to accomodate the
addition of two new crates; all of the previous work has now been moved
into 'compiler' - in the future the name for this directory might
change, but that will come after a name for the compiler can be nailed
down.